fixed build errors

+++ includes/constants.h
@@ -81,8 +81,8 @@ namespace BOARD {
 
 namespace TIME {
 
-	const int PASSIVE_TIME_CUTOFF = 1.f;
+	const float PASSIVE_TIME_CUTOFF = 1.f;
 
-	const int PIECE_MOVE_COOLDOWN = .2f;
+	const float PIECE_MOVE_COOLDOWN = .2f;
 
  }
